+++ doc/conf.py
@@ -8,3 +8,5 @@ automodsumm_inherited_members = True  # noqa
 autodoc_inherit_docstrings = True  # noqa
 autoclass_content = "class"  # noqa
 autodoc_default_flags = ["show-inheritance", "special-members"]  # noqa
+extensions.remove("myst_parser")  # noqa
+source_suffix = {".rst": "restructuredtext"}